auth ! JWt backend

+++ backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 // @desc Get goals
 // @route GET /api/goals
 // @acces private
 const getGoals = asyncHandler(async (req, res) => {
-	const goals = await Goal.find();
+	const goals = await Goal.find({ user: req.user.id });
 
 	res.status(200).json(goals);
 });
@@ -22,6 +23,7 @@ const setGoals = asyncHandler(async (req, res) => {
 
 	const goal = await Goal.create({
 		text: req.body.text,
+		user: req.user.id,
 	});
 	res.status(200).json(goal);
 });
@@ -36,6 +38,20 @@ const updateGoal = asyncHandler(async (req, res) => {
 		throw new Error('goal not found');
 	}
 
+	const user = await User.findById(req.user.id);
+
+	//Check for user
+	if (!user) {
+		res.status(401);
+		throw new Error('user not found');
+	}
+
+	// Make sure the loggged in user matches the goal user
+
+	if (goal.user.toString() !== user.id) {
+		res.status(401);
+		throw new Error('user not authorized');
+	}
 	const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
 		new: true,
 	});
@@ -52,6 +68,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
 	if (!goal) {
 		res.status(400);
 		throw new Error('goal not found');
+	}
+
+	const user = await User.findById(req.user.id);
+
+	//Check for user
+	if (!user) {
+		res.status(401);
+		throw new Error('user not found');
+	}
+
+	// Make sure the loggged in user matches the goal user
+
+	if (goal.user.toString() !== user.id) {
+		res.status(401);
+		throw new Error('user not authorized');
 	}
 	await goal.deleteOne();
 	res.status(200).json({ id: req.params.id });
